Convert rdkit molecule to Topology
